+++ src/main/java/org/opentripplanner/transit/raptor/api/transit/RaptorRoute.java
@@ -32,10 +32,10 @@ public interface RaptorRoute<T extends RaptorTripSchedule> {
     /**
      * Return a guaranteed transfer provider for the given route transferring TO the route.
      */
-    RaptorForbiddenTransferProvider<T> getForbiddenTransfersTo();
+    RaptorForbiddenStopTransferProvider<T> getForbiddenTransfersTo();
 
     /**
      * Return a guaranteed transfer provider for the given route transferring FROM the route.
      */
-    RaptorForbiddenTransferProvider<T> getForbiddenTransfersFrom();
+    RaptorForbiddenStopTransferProvider<T> getForbiddenTransfersFrom();
 }
