+++ src/main/java/org/opentripplanner/transit/raptor/rangeraptor/transit/TransitCalculator.java
@@ -12,7 +12,7 @@ import org.opentripplanner.transit.raptor.api.transit.RaptorTransfer;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
 
 
-import org.opentripplanner.transit.raptor.api.transit.RaptorForbiddenTransferProvider;
+import org.opentripplanner.transit.raptor.api.transit.RaptorForbiddenStopTransferProvider;
 
 /**
  * The transit calculator is used to calculate transit related stuff, like calculating
@@ -165,7 +165,7 @@ public interface TransitCalculator<T extends RaptorTripSchedule> {
      * target is the FROM pattern/stop.
      */
     RaptorGuaranteedTransferProvider<T> guaranteedTransfers(RaptorRoute<T> route);
-    RaptorForbiddenTransferProvider<T> forbiddenTransfers(RaptorRoute<T> route);
+    RaptorForbiddenStopTransferProvider<T> forbiddenTransfers(RaptorRoute<T> route);
 
     /**
      * Return a calculator for test purpose. The following parameters are fixed:
